+++ examples/simple/extism/main.go
@@ -23,6 +23,7 @@ func runExtismExample(logger *slog.Logger) (map[string]any, error) {
 	}
 
 	evaluator, err := polyscript.New[polyscript.Extism](
+		context.Background(),
 		polyscript.FromBytes(wasmdata.TestModule),
 		polyscript.WithEntryPoint(wasmdata.EntrypointGreet),
 		polyscript.WithStaticData[polyscript.Extism](inputData),